requirements.txt
pytz
requests
python-telegram-bot
